Add the majority of responsive hero functionality
This includes fixing where the image sits on overly wide screens and
fixing the wrapping of the linked buttons and header text, and, on
mobile screens, rotating the image, removing the logo, shrinking
the text size, and stacking the elements. Ongoing problems include
text overlaps with image and overlay doesn't always stay at edge
of the image on xsmall screens.