Fix bolded text issue on Chrome & bold text cleanup
This adds an import for the weight 600 version of the Poppins font,
which fixes an issue where bolded text was not showing up as
expected on Chrome. This commit also cleans up the places where
bold text should appear, including changing the hover state
on the header links to match the designs.